+++ utils/buffer/reader.go
@@ -77,8 +77,13 @@ func ReadUint16Slice(r Reader, c []uint16) (n int, err error) {
 
 	var slice []byte
 
+	size := r.Size()
+	if len(c)<<1 < size {
+		size = len(c) << 1
+	}
+
 	// Then returns the unread bytes
-	if slice, err = r.Peek(r.Size()); err != nil {
+	if slice, err = r.Peek(size); err != nil {
 		fmt.Println(err)
 		return
 	}
@@ -143,8 +148,14 @@ func ReadUint32Slice(r Reader, c []uint32) (n int, err error) {
 
 	var slice []byte
 
+	// Avoid EOF
+	size := r.Size()
+	if len(c)<<2 < size {
+		size = len(c) << 2
+	}
+
 	// Then returns the unread bytes
-	if slice, err = r.Peek(r.Size()); err != nil {
+	if slice, err = r.Peek(size); err != nil {
 		fmt.Println(err)
 		return
 	}
@@ -209,9 +220,14 @@ func ReadUint64Slice(r Reader, c []uint64) (n int, err error) {
 
 	var slice []byte
 
+	// Avoid EOF
+	size := r.Size()
+	if len(c)<<3 < size {
+		size = len(c) << 3
+	}
+
 	// Then returns the unread bytes
-	if slice, err = r.Peek(r.Size()); err != nil {
-		fmt.Println(err)
+	if slice, err = r.Peek(size); err != nil {
 		return
 	}